pull in PlaneStompsWideGlyph from render.cc
